requirements.txt
Django==4.1.3
pymongo==4.3.3
python-dotenv~=0.21.0